create reconfiguraion_kvbc_handler.cpp
